Still trying to get popup in working order.
