+++ Incorpreal/DialogueBox.tscn
@@ -1,0 +1,100 @@
+[gd_scene load_steps=3 format=2]
+
+[ext_resource path="res://TextPopup.cs" type="Script" id=1]
+
+[sub_resource type="Animation" id=1]
+resource_name = "ShowDialogue"
+length = 1.1
+tracks/0/type = "value"
+tracks/0/path = NodePath("ColorRect/ObjectText:percent_visible")
+tracks/0/interp = 1
+tracks/0/loop_wrap = true
+tracks/0/imported = false
+tracks/0/enabled = true
+tracks/0/keys = {
+"times": PoolRealArray( 0.2, 0.9 ),
+"transitions": PoolRealArray( 1, 1 ),
+"update": 0,
+"values": [ 0.0, 1.0 ]
+}
+tracks/1/type = "value"
+tracks/1/path = NodePath("ColorRect/PlayerOptions:visible")
+tracks/1/interp = 1
+tracks/1/loop_wrap = true
+tracks/1/imported = false
+tracks/1/enabled = true
+tracks/1/keys = {
+"times": PoolRealArray( 0.1, 1.1 ),
+"transitions": PoolRealArray( 1, 1 ),
+"update": 1,
+"values": [ false, true ]
+}
+
+[node name="Control" type="Control"]
+margin_right = 40.0
+margin_bottom = 40.0
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="TextPopup" type="Popup" parent="."]
+visible = true
+anchor_left = 0.5
+anchor_top = 1.0
+anchor_right = 0.5
+anchor_bottom = 1.0
+margin_left = -20.0
+margin_top = 320.0
+margin_right = 620.0
+margin_bottom = 320.0
+popup_exclusive = true
+script = ExtResource( 1 )
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="ColorRect" type="ColorRect" parent="TextPopup"]
+margin_left = 10.0
+margin_top = 115.0
+margin_right = 310.0
+margin_bottom = 175.0
+
+[node name="ObjectName" type="Label" parent="TextPopup/ColorRect"]
+margin_left = 5.0
+margin_top = 2.0
+margin_right = 295.0
+margin_bottom = 16.0
+custom_colors/font_color = Color( 0, 0, 0, 1 )
+text = "Fiona"
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="ObjectText" type="Label" parent="TextPopup/ColorRect"]
+margin_left = 5.0
+margin_top = 15.0
+margin_right = 295.0
+margin_bottom = 40.0
+custom_colors/font_color = Color( 0.321569, 0.313726, 0.313726, 1 )
+text = "Hello adventurer! I lost my necklace, can you find it for me?"
+autowrap = true
+percent_visible = 0.0
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="PlayerOptions" type="Label" parent="TextPopup/ColorRect"]
+margin_left = 5.0
+margin_top = 45.0
+margin_right = 295.0
+margin_bottom = 59.0
+custom_colors/font_color = Color( 0, 0, 0, 1 )
+text = "[Enter] Yes [Tab] No"
+align = 1
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="DialogueAnimation" type="AnimationPlayer" parent="TextPopup"]
+anims/ShowDialogue = SubResource( 1 )
+[connection signal="animation_finished" from="TextPopup/DialogueAnimation" to="TextPopup" method="_on_DialogueAnimation_animation_finished"]
